Factory architecture and componentization

+++ test/flow_concret.cpp
@@ -4,11 +4,11 @@
 
 #include "flow_concret.h"
 
-float FlowExp::execute() {
+double FlowExp::execute() {
     return getSource()->getValue() * 0.01;
 }
 
-float FlowLog::execute() {
+double FlowLog::execute() {
     return 0.01 * getTarget()->getValue() * (1 - getTarget()->getValue() / 70);
 
 }
